+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.3.0] - 2019-10-22
 ### Added
 - Add `pause` and `resume` on `PlayAction`.
 - Add `volume` optional parameter to `play` and `prompt` methods